refactor(Timer): use jsx and unneeded components
