+++ src/components/Timer/Selections.js
@@ -1,5 +1,5 @@
 import React from 'react';
-import './Timer.css';
+import './index.css';
 import { TimerState } from 'const/Timer';
 import AppBar from '@material-ui/core/AppBar';
 import Tabs from '@material-ui/core/Tabs';
