+++ src/components/Timer/index.css
