--- src/components/timer/Avocombo.js
@@ -1,7 +1,0 @@
-import React from 'react';
-
-const Avocombo = props => {
-  return <div>{props.avocomboCount}</div>;
-};
-
-export default Avocombo;
