--- src/components/timer/TimerInfo.js
@@ -1,7 +1,0 @@
-import React from 'react';
-
-const TimerInfo = props => {
-  return <div>{props.getTime}</div>;
-};
-
-export default TimerInfo;
