+++ src/const/Timer.js
@@ -1,4 +1,4 @@
-const AVOCADORO_TIME = 1500;
+const AVOCADORO_TIME = 10;
 const SHORT_BREAK_TIME = 300;
 const LONG_BREAK_TIME = 900;
