+++ src/containers/Timer.js
@@ -1,5 +1,5 @@
 import { connect } from 'react-redux';
-import Timer from 'components/timer/Timer';
+import Timer from 'components/Timer';
 
 const mapStateToProps = (state, ownProps) => {
   return {
